byte

+++ pacman-core/src/main/java/de/amr/games/pacman/model/common/GameModel.java
@@ -181,7 +181,7 @@ public abstract class GameModel {
 
 	protected int[] globalDotLimits;
 
-	protected int[] privateDotLimits;
+	protected byte[] privateDotLimits;
 
 	/** Number of current intermission scene in test mode. */
 	public int intermissionTestNumber;
