+++ pacman-core/src/main/java/de/amr/games/pacman/model/pacman/PacManGame.java
@@ -188,9 +188,9 @@ public class PacManGame extends GameModel {
 		pacStarvingTimeLimit = (int) secToTicks(level.number() < 5 ? 4 : 3);
 		globalDotLimits = new int[] { Integer.MAX_VALUE, 7, 17, Integer.MAX_VALUE };
 		privateDotLimits = switch (level.number()) {
-		case 1 -> new int[] { 0, 0, 30, 60 };
-		case 2 -> new int[] { 0, 0, 0, 50 };
-		default -> new int[] { 0, 0, 0, 0 };
+		case 1 -> new byte[] { 0, 0, 30, 60 };
+		case 2 -> new byte[] { 0, 0, 0, 50 };
+		default -> new byte[] { 0, 0, 0, 0 };
 		};
 		numGhostsKilledInLevel = 0;
 	}
